Import of MLEHMANN/Video-Capture-V4l-0.902 from CPAN.
gitpan-cpan-distribution: Video-Capture-V4l
gitpan-cpan-version:      0.902
gitpan-cpan-path:         MLEHMANN/Video-Capture-V4l-0.902.tar.gz
gitpan-cpan-author:       MLEHMANN
gitpan-cpan-maturity:     released

+++ RTjpeg/codec/modules/RTb2s.c
@@ -2,7 +2,7 @@ int RTjpeg_b2s(s16 *data, s8 *strm, u8 bt8)
 {
  int ci=1, co=1, tmp;
 
- (u8)strm[0]=(u8)(data[RTjpeg_ZZ[0]]>254) ? 254:((data[RTjpeg_ZZ[0]]<0)?0:data[RTjpeg_ZZ[0]]);
+ strm[0]=(u8)(data[RTjpeg_ZZ[0]]>254) ? 254:((data[RTjpeg_ZZ[0]]<0)?0:data[RTjpeg_ZZ[0]]);
  
  for(ci=1; ci<=bt8; ci++)
   if(data[RTjpeg_ZZ[ci]]>0)
